Crate rustc_hir_typeck

Modules

_match 🔒

autoderef 🔒

Some helper functions for AutoDeref.

callee 🔒

cast

Code for type-checking cast expressions.

check 🔒

closure 🔒

Code for type-checking closure expressions.

coercion 🔒

Type Coercion

demand 🔒

diverges 🔒

errors 🔒

Errors emitted by rustc_hir_typeck.

expectation 🔒

expr 🔒

Type checking expressions.

expr_use_visitor

A different sort of visitor for walking fn bodies. Unlike the normal visitor, which just walks the entire body in one shot, the ExprUseVisitor determines how expressions are being used.

fallback 🔒

fn_ctxt 🔒

gather_locals 🔒

inline_asm 🔒

intrinsicck 🔒

loops 🔒

method 🔒

Method lookup: the secret sauce of Rust. See the rustc dev guide for more information.

naked_functions 🔒

Checks validity of naked functions.

op 🔒

Code related to processing overloaded binary and unary operators.

opaque_types 🔒

pat 🔒

place_op 🔒

rvalue_scopes 🔒

typeck_root_ctxt 🔒

upvar 🔒

Inferring borrow kinds for upvars

writeback 🔒

During type inference, partially inferred terms are represented using inference variables (ty::Infer). These don’t appear in the final ty::TypeckResults since all of the types should have been inferred once typeck is done.

Macros

type_error_struct

Structs

BreakableCtxt

CoroutineTypes 🔒

When check_fn is invoked on a coroutine (i.e., a body that includes yield), it returns back some information about the yield points.

EnclosingBreakables

Enums

Needs

PlaceOp

TupleArgumentsFlag 🔒

Controls whether the arguments are tupled. This is used for the call operator.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

can_coerce

Check whether ty can be coerced to output_ty. Used from clippy.

fatally_break_rust 🔒

infer_type_if_missing 🔒

inspect_typeck

Same as typeck but inspect is invoked on evaluation of each root obligation. Inspecting obligations only works with the new trait solver. This function is only to be used by external tools, it should not be called from within rustc. Note, this is not a query, and thus is not cached.

provide

Adds query implementations to the Providers vtable, see rustc_middle::query

report_unexpected_variant_res 🔒

typeck 🔒

typeck_with_inspect 🔒

used_trait_imports 🔒