Struct rustc_hir_typeck::cast::CastCheck

pub struct CastCheck<'tcx> {
    expr: &'tcx Expr<'tcx>,
    expr_ty: Ty<'tcx>,
    expr_span: Span,
    cast_ty: Ty<'tcx>,
    cast_span: Span,
    span: Span,
    pub constness: Constness,
}

Reifies a cast check to be checked once we have full type information for a function context.

Fields
expr: &'tcx Expr<'tcx>

The expression whose value is being casted

expr_ty: Ty<'tcx>

The source type for the cast expression

expr_span: Spancast_ty: Ty<'tcx>

The target type. That is, the type we are casting to.

cast_span: Spanspan: Spanconstness: Constness

whether the cast is made in a const context or not.

Implementations

impl<'a, 'tcx> CastCheck<'tcx>

pub fn new( fcx: &FnCtxt<'a, 'tcx>, expr: &'tcx Expr<'tcx>, expr_ty: Ty<'tcx>, cast_ty: Ty<'tcx>, cast_span: Span, span: Span, constness: Constness ) -> Result<CastCheck<'tcx>, ErrorGuaranteed>

fn report_cast_error(&self, fcx: &FnCtxt<'a, 'tcx>, e: CastError)

fn report_cast_to_unsized_type(&self, fcx: &FnCtxt<'a, 'tcx>) -> ErrorGuaranteed

fn trivial_cast_lint(&self, fcx: &FnCtxt<'a, 'tcx>)

pub fn check(self, fcx: &FnCtxt<'a, 'tcx>)

pub fn do_check(&self, fcx: &FnCtxt<'a, 'tcx>) -> Result<CastKind, CastError>

Checks a cast, and report an error if one exists. In some cases, this can return Ok and create type errors in the fcx rather than returning directly. coercion-cast is handled in check instead of here.

fn check_ptr_ptr_cast( &self, fcx: &FnCtxt<'a, 'tcx>, m_expr: TypeAndMut<'tcx>, m_cast: TypeAndMut<'tcx> ) -> Result<CastKind, CastError>

fn check_fptr_ptr_cast( &self, fcx: &FnCtxt<'a, 'tcx>, m_cast: TypeAndMut<'tcx> ) -> Result<CastKind, CastError>

fn check_ptr_addr_cast( &self, fcx: &FnCtxt<'a, 'tcx>, m_expr: TypeAndMut<'tcx> ) -> Result<CastKind, CastError>

fn check_ref_cast( &self, fcx: &FnCtxt<'a, 'tcx>, m_expr: TypeAndMut<'tcx>, m_cast: TypeAndMut<'tcx> ) -> Result<CastKind, CastError>

fn check_addr_ptr_cast( &self, fcx: &FnCtxt<'a, 'tcx>, m_cast: TypeAndMut<'tcx> ) -> Result<CastKind, CastError>

fn try_coercion_cast( &self, fcx: &FnCtxt<'a, 'tcx> ) -> Result<(), TypeError<'tcx>>

fn cenum_impl_drop_lint(&self, fcx: &FnCtxt<'a, 'tcx>)

fn lossy_provenance_ptr2int_lint(&self, fcx: &FnCtxt<'a, 'tcx>, t_c: IntTy)

fn fuzzy_provenance_int2ptr_lint(&self, fcx: &FnCtxt<'a, 'tcx>)

fn try_suggest_collection_to_bool( &self, fcx: &FnCtxt<'a, 'tcx>, err: &mut Diagnostic )

Attempt to suggest using .is_empty when trying to cast from a collection type to a boolean.