Struct rustc_hir_typeck::coercion::Coerce

struct Coerce<'a, 'tcx> {
    fcx: &'a FnCtxt<'a, 'tcx>,
    cause: ObligationCause<'tcx>,
    use_lub: bool,
    allow_two_phase: AllowTwoPhase,
}

Fields
fcx: &'a FnCtxt<'a, 'tcx>cause: ObligationCause<'tcx>use_lub: boolallow_two_phase: AllowTwoPhase

Determines whether or not allow_two_phase_borrow is set on any autoref adjustments we create while coercing. We don’t want to allow deref coercions to create two-phase borrows, at least initially, but we do need two-phase borrows for function argument reborrows. See #47489 and #48598 See docs on the “AllowTwoPhase” type for a more detailed discussion

Implementations

impl<'f, 'tcx> Coerce<'f, 'tcx>

fn new(
fcx: &'f FnCtxt<'f, 'tcx>,
cause: ObligationCause<'tcx>,
allow_two_phase: AllowTwoPhase
) -> Self

fn unify(&self, a: Ty<'tcx>, b: Ty<'tcx>) -> InferResult<'tcx, Ty<'tcx>>

fn unify_and<F>(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>,
f: F
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>where
F: FnOnce(Ty<'tcx>) -> Vec<Adjustment<'tcx>>,

Unify two types (using sub or lub) and produce a specific coercion.

fn coerce(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

fn coerce_from_inference_variable(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>,
make_adjustments: impl FnOnce(Ty<'tcx>) -> Vec<Adjustment<'tcx>>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

Coercing from an inference variable. In this case, we have no information about the source type, so we can’t really do a true coercion and we always fall back to subtyping (unify_and).

fn coerce_borrowed_pointer(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>,
r_b: Region<'tcx>,
mutbl_b: Mutability
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

Reborrows &mut A to &mut B and &(mut) A to &B. To match A with B, autoderef will be performed, calling deref/deref_mut where necessary.

fn coerce_unsized(
&self,
source: Ty<'tcx>,
target: Ty<'tcx>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

fn coerce_dyn_star(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>,
predicates: &'tcx List<PolyExistentialPredicate<'tcx>>,
b_region: Region<'tcx>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

fn coerce_from_safe_fn<F, G>(
&self,
a: Ty<'tcx>,
fn_ty_a: PolyFnSig<'tcx>,
b: Ty<'tcx>,
to_unsafe: F,
normal: G
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>where
F: FnOnce(Ty<'tcx>) -> Vec<Adjustment<'tcx>>,
G: FnOnce(Ty<'tcx>) -> Vec<Adjustment<'tcx>>,

fn coerce_from_fn_pointer(
&self,
a: Ty<'tcx>,
fn_ty_a: PolyFnSig<'tcx>,
b: Ty<'tcx>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

Attempts to coerce from the type of a Rust function item into a closure or a proc.

fn coerce_from_fn_item(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

Attempts to coerce from the type of a Rust function item into a closure or a proc.

fn coerce_closure_to_fn(
&self,
a: Ty<'tcx>,
closure_def_id_a: DefId,
substs_a: SubstsRef<'tcx>,
b: Ty<'tcx>
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>

Attempts to coerce from the type of a non-capturing closure into a function pointer.

fn coerce_unsafe_ptr(
&self,
a: Ty<'tcx>,
b: Ty<'tcx>,
mutbl_b: Mutability
) -> InferResult<'tcx, (Vec<Adjustment<'tcx>>, Ty<'tcx>)>