Module rustc_hir_typeck::expr_use_visitor

A different sort of visitor for walking fn bodies. Unlike the normal visitor, which just walks the entire body in one shot, the ExprUseVisitor determines how expressions are being used.

Macros

return_if_err 🔒

If the MC results in an error, it’s because the type check failed (or will fail, when the error is uncovered and reported during writeback). In this case, we just ignore this part of the code.

Structs

ExprUseVisitor

The ExprUseVisitor type

Place

A Place represents how a value is located in memory.

PlaceWithHirId

A PlaceWithHirId represents how a value is located in memory.

Projection

Enums

ConsumeMode 🔒

PlaceBase

Traits

Delegate

This trait defines the callbacks you can expect to receive when employing the ExprUseVisitor.

Functions

copy_or_move 🔒

delegate_consume 🔒

is_multivariant_adt 🔒