rustc_hir_typeck

Function inspect_typeck

pub fn inspect_typeck<'tcx>(
    tcx: TyCtxt<'tcx>,
    def_id: LocalDefId,
    inspect: ObligationInspector<'tcx>,
) -> &'tcx TypeckResults<'tcx>

Same as typeck but inspect is invoked on evaluation of each root obligation. Inspecting obligations only works with the new trait solver. This function is only to be used by external tools, it should not be called from within rustc. Note, this is not a query, and thus is not cached.