Struct rustc_hir_typeck::fn_ctxt::arg_matrix::ArgMatrix

pub(crate) struct ArgMatrix<'tcx> {
    provided_indices: Vec<ProvidedIdx>,
    expected_indices: Vec<ExpectedIdx>,
    compatibility_matrix: Vec<Vec<Compatibility<'tcx>>>,
}

Fields
provided_indices: Vec<ProvidedIdx>

Maps the indices in the compatibility_matrix rows to the indices of the user provided inputs

expected_indices: Vec<ExpectedIdx>

Maps the indices in the compatibility_matrix columns to the indices of the expected args

compatibility_matrix: Vec<Vec<Compatibility<'tcx>>>

The first dimension (rows) are the remaining user provided inputs to match and the second dimension (cols) are the remaining expected args to match

Implementations

impl<'tcx> ArgMatrix<'tcx>

pub(crate) fn new<F: FnMut(ProvidedIdx, ExpectedIdx) -> Compatibility<'tcx>>( provided_count: usize, expected_input_count: usize, is_compatible: F ) -> Self

fn eliminate_provided(&mut self, idx: usize)

Remove a given input from consideration

fn eliminate_expected(&mut self, idx: usize)

Remove a given argument from consideration

fn satisfy_input(&mut self, provided_idx: usize, expected_idx: usize)

“satisfy” an input with a given arg, removing both from consideration

fn eliminate_satisfied(&mut self) -> Vec<(ProvidedIdx, ExpectedIdx)>

fn find_issue(&self) -> Option<Issue>

pub(crate) fn find_errors( self ) -> (Vec<Error<'tcx>>, IndexVec<ExpectedIdx, Option<ProvidedIdx>>)