Struct rustc_hir_typeck::method::confirm::ConfirmContext

struct ConfirmContext<'a, 'tcx> {
    fcx: &'a FnCtxt<'a, 'tcx>,
    span: Span,
    self_expr: &'tcx Expr<'tcx>,
    call_expr: &'tcx Expr<'tcx>,
    skip_record_for_diagnostics: bool,
}

Fields
fcx: &'a FnCtxt<'a, 'tcx>span: Spanself_expr: &'tcx Expr<'tcx>call_expr: &'tcx Expr<'tcx>skip_record_for_diagnostics: bool
Implementations

impl<'a, 'tcx> ConfirmContext<'a, 'tcx>

fn new( fcx: &'a FnCtxt<'a, 'tcx>, span: Span, self_expr: &'tcx Expr<'tcx>, call_expr: &'tcx Expr<'tcx> ) -> ConfirmContext<'a, 'tcx>

fn confirm( &mut self, unadjusted_self_ty: Ty<'tcx>, pick: &Pick<'tcx>, segment: &PathSegment<'_> ) -> ConfirmResult<'tcx>

fn adjust_self_ty( &mut self, unadjusted_self_ty: Ty<'tcx>, pick: &Pick<'tcx> ) -> Ty<'tcx>

fn fresh_receiver_args( &mut self, self_ty: Ty<'tcx>, pick: &Pick<'tcx> ) -> GenericArgsRef<'tcx>

Returns a set of substitutions for the method receiver where all type and region parameters are instantiated with fresh variables. This substitution does not include any parameters declared on the method itself.

Note that this substitution may include late-bound regions from the impl level. If so, these are instantiated later in the instantiate_method_sig routine.

fn extract_existential_trait_ref<R, F>( &mut self, self_ty: Ty<'tcx>, closure: F ) -> Rwhere F: FnMut(&mut ConfirmContext<'a, 'tcx>, Ty<'tcx>, PolyExistentialTraitRef<'tcx>) -> R,

fn instantiate_method_args( &mut self, pick: &Pick<'tcx>, seg: &PathSegment<'_>, parent_args: GenericArgsRef<'tcx> ) -> GenericArgsRef<'tcx>

fn unify_receivers( &mut self, self_ty: Ty<'tcx>, method_self_ty: Ty<'tcx>, pick: &Pick<'tcx>, args: GenericArgsRef<'tcx> )

fn instantiate_method_sig( &mut self, pick: &Pick<'tcx>, all_args: GenericArgsRef<'tcx> ) -> (FnSig<'tcx>, InstantiatedPredicates<'tcx>)

fn add_obligations( &mut self, fty: Ty<'tcx>, all_args: GenericArgsRef<'tcx>, method_predicates: InstantiatedPredicates<'tcx>, def_id: DefId )

fn predicates_require_illegal_sized_bound( &self, predicates: InstantiatedPredicates<'tcx> ) -> Option<Span>

fn enforce_illegal_method_limitations(&self, pick: &Pick<'_>)

fn upcast( &mut self, source_trait_ref: PolyTraitRef<'tcx>, target_trait_def_id: DefId ) -> PolyTraitRef<'tcx>

fn instantiate_binder_with_fresh_vars<T>(&self, value: Binder<'tcx, T>) -> Twhere T: TypeFoldable<TyCtxt<'tcx>> + Copy,