Struct rustc_hir_typeck::upvar::InferBorrowKind

struct InferBorrowKind<'a, 'tcx> {
    fcx: &'a FnCtxt<'a, 'tcx>,
    closure_def_id: LocalDefId,
    capture_information: Vec<(Place<'tcx>, CaptureInfo)>,
    fake_reads: Vec<(Place<'tcx>, FakeReadCause, HirId)>,
}

Fields
fcx: &'a FnCtxt<'a, 'tcx>closure_def_id: LocalDefIdcapture_information: Vec<(Place<'tcx>, CaptureInfo)>

For each Place that is captured by the closure, we track the minimal kind of access we need (ref, ref mut, move, etc) and the expression that resulted in such access.

Consider closure where s.str1 is captured via an ImmutableBorrow and s.str2 via a MutableBorrow

struct SomeStruct { str1: String, str2: String };

// Assume that the HirId for the variable definition is `V1`
let mut s = SomeStruct { str1: format!("s1"), str2: format!("s2") };

let fix_s = |new_s2| {
    // Assume that the HirId for the expression `s.str1` is `E1`
    println!("Updating SomeStruct with str1={0}", s.str1);
    // Assume that the HirId for the expression `*s.str2` is `E2`
    s.str2 = new_s2;
};

For closure fix_s, (at a high level) the map contains

Place { V1, [ProjectionKind::Field(Index=0, Variant=0)] } : CaptureKind { E1, ImmutableBorrow }
Place { V1, [ProjectionKind::Field(Index=1, Variant=0)] } : CaptureKind { E2, MutableBorrow }

fake_reads: Vec<(Place<'tcx>, FakeReadCause, HirId)>