rustc_hir_typeck::coercion
Type Alias DynamicCoerceMany

pub(crate) type DynamicCoerceMany<'tcx> = CoerceMany<'tcx, 'tcx, &'tcx Expr<'tcx>>;

The type of a CoerceMany that is storing up the expressions into a buffer. We use this in check/mod.rs for things like break.

Aliased Type

struct DynamicCoerceMany<'tcx> {
    expected_ty: Ty<'tcx>,
    final_ty: Option<Ty<'tcx>>,
    expressions: Expressions<'tcx, 'tcx, &'tcx Expr<'tcx>>,
    pushed: usize,
}

Fields
expected_ty: Ty<'tcx>final_ty: Option<Ty<'tcx>>expressions: Expressions<'tcx, 'tcx, &'tcx Expr<'tcx>>pushed: usize
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 48 bytes