rustc_hir_typeck
Module expr_use_visitor

A different sort of visitor for walking fn bodies. Unlike the normal visitor, which just walks the entire body in one shot, the ExprUseVisitor determines how expressions are being used.

Structs

ExprUseVisitor

The ExprUseVisitor type

Place

A Place represents how a value is located in memory.

PlaceWithHirId

A PlaceWithHirId represents how a value is located in memory.

Projection

Enums

PlaceBase

Traits

Delegate

This trait defines the callbacks you can expect to receive when employing the ExprUseVisitor.

TypeInformationCtxt