rustc_hir_typeck::method
Module probe

Structs

Candidate 🔒

IsSuggestion 🔒

Boolean flag used to indicate if this search is for a suggestion or not. If true, we can allow ambiguity and so forth.

Pick 🔒

PickConstraintsForShadowed 🔒

Criteria to apply when searching for a given Pick. This is used during the search for potentially shadowed methods to ensure we don’t search more candidates than strictly necessary.

PickDiagHints 🔒

Extra information required only for error reporting.

ProbeContext 🔒

Enums

AutorefOrPtrAdjustment 🔒

When adjusting a receiver we often want to do one of

CandidateKind 🔒

Mode 🔒

PickKind 🔒

ProbeResult 🔒

ProbeScope 🔒

Functions

method_autoderef_steps 🔒

provide 🔒

Type Aliases

PickResult 🔒