Module rustc_hir_typeck::method::probe

Re-exports

pub use self::PickKind::*;

Structs

Candidate 🔒

IsSuggestion

Boolean flag used to indicate if this search is for a suggestion or not. If true, we can allow ambiguity and so forth.

Pick

ProbeContext 🔒

Enums

AutorefOrPtrAdjustment

When adjusting a receiver we often want to do one of

CandidateKind 🔒

Mode

PickKind

ProbeResult 🔒

ProbeScope

Functions

method_autoderef_steps 🔒

provide

Type Aliases

PickResult