rustc_hir_typeck::method
Module suggest

Give useful errors and suggestions to users when an item can’t be found or is otherwise invalid.

Structs

TraitInfo 🔒

Enums

SelfSource 🔒

Functions

all_traits 🔒

Retrieves all traits in this crate and any dependent crates, and wraps them into TraitInfo for custom sorting.

print_disambiguation_help 🔒