Module rustc_hir_typeck::pat

Structs

PatInfo 🔒

TopInfo 🔒

Information about the expected type at the top level of type checking a pattern.

Enums

AdjustMode 🔒

Mode for adjusting the expected type and binding mode.

MutblCap 🔒

ref mut patterns (explicit or match-ergonomics) are not allowed behind an & reference.

Constants

CANNOT_IMPLICITLY_DEREF_POINTER_TRAIT_OBJ 🔒

Functions

is_number 🔒