Function rustc_hir_typeck::upvar::restrict_capture_precision

fn restrict_capture_precision(
    place: Place<'_>,
    curr_mode: UpvarCapture,
) -> (Place<'_>, UpvarCapture)

Truncate projections so that following rules are obeyed by the captured place:

No Index projections are captured, since arrays are captured completely.

No unsafe block is required to capture place Returns the truncated place and updated capture mode.